Moved the tests suites for the different emitters into their own test folders
